Update README.md
This commit adds or updates the README file for the ASL Recognition with Deep Learning project. The README file provides an overview of the project, including information about the datasets, dependencies, and getting started. It also includes a license for the project. The goal of the README file is to provide clear and concise instructions for using the project and to help users understand the purpose and functionality of the code.